clean up irc catalog code
