+++ src/curl.cpp
@@ -85,7 +85,7 @@ public:
 };
 
 static size_t RequestWriteCallback(void *contents, size_t size, size_t nmemb, void *userp) {
-	((std::string *)userp)->append((char *)contents, size * nmemb);
+	((string *)userp)->append((char *)contents, size * nmemb);
 	return size * nmemb;
 }
