+++ src/include/storage/irc_catalog.hpp
@@ -10,24 +10,20 @@
 #include "rest_catalog/objects/load_table_result.hpp"
 #include "storage/irc_authorization.hpp"
 
+#include "duckdb/parser/parsed_data/attach_info.hpp"
+#include "duckdb/storage/storage_extension.hpp"
+
 namespace duckdb {
 
 class IRCSchemaEntry;
 
-class ICRClearCacheFunction : public TableFunction {
-public:
-	ICRClearCacheFunction();
-
-	static void ClearCacheOnSetting(ClientContext &context, SetScope scope, Value &parameter);
-};
-
 class MetadataCacheValue {
 public:
-	const std::string data;
+	const string data;
 	const system_clock::time_point expires_at;
 
 public:
-	MetadataCacheValue(const std::string &data_, const system_clock::time_point expires_at_)
+	MetadataCacheValue(const string &data_, const system_clock::time_point expires_at_)
 	    : data(data_), expires_at(expires_at_) {
 	}
 };
@@ -38,44 +34,32 @@ public:
 	                   const string &warehouse, const string &uri, const string &version = "v1");
 	~IRCatalog() override;
 
-	string internal_name;
-	AccessMode access_mode;
-	unique_ptr<IRCAuthorization> auth_handler;
-	IRCEndpointBuilder endpoint_builder;
+public:
+	static unique_ptr<SecretEntry> GetStorageSecret(ClientContext &context, const string &secret_name);
+	static unique_ptr<SecretEntry> GetIcebergSecret(ClientContext &context, const string &secret_name);
+	void GetConfig(ClientContext &context);
+	IRCEndpointBuilder GetBaseUrl() const;
+	string OptionalGetCachedValue(const string &url);
+	bool SetCachedValue(const string &url, const string &value, const rest_api_objects::LoadTableResult &result);
+	string DefaultSchema() const;
 
-	//! warehouse
-	string warehouse;
-
-	//! host of the REST catalog
-	string uri;
-	//! version
-	const string version;
-	//! optional prefix
-	string prefix;
+public:
+	static unique_ptr<Catalog> Attach(StorageExtensionInfo *storage_info, ClientContext &context, AttachedDatabase &db,
+	                                  const string &name, AttachInfo &info, AccessMode access_mode);
 
 public:
 	void Initialize(bool load_builtin) override;
-
 	string GetCatalogType() override {
 		return "iceberg";
 	}
-
-	static unique_ptr<SecretEntry> GetStorageSecret(ClientContext &context, const string &secret_name);
-	static unique_ptr<SecretEntry> GetIcebergSecret(ClientContext &context, const string &secret_name);
-
-	void GetConfig(ClientContext &context);
-
 	bool SupportsTimeTravel() const override {
 		return true;
 	}
-
+	void DropSchema(ClientContext &context, DropInfo &info) override;
 	optional_ptr<CatalogEntry> CreateSchema(CatalogTransaction transaction, CreateSchemaInfo &info) override;
-
 	void ScanSchemas(ClientContext &context, std::function<void(SchemaCatalogEntry &)> callback) override;
-
 	optional_ptr<SchemaCatalogEntry> LookupSchema(CatalogTransaction transaction, const EntryLookupInfo &schema_lookup,
 	                                              OnEntryNotFound if_not_found) override;
-
 	PhysicalOperator &PlanInsert(ClientContext &context, PhysicalPlanGenerator &planner, LogicalInsert &op,
 	                             optional_ptr<PhysicalOperator> plan) override;
 	PhysicalOperator &PlanCreateTableAs(ClientContext &context, PhysicalPlanGenerator &planner, LogicalCreateTable &op,
@@ -86,26 +70,26 @@ public:
 	                             PhysicalOperator &plan) override;
 	unique_ptr<LogicalOperator> BindCreateIndex(Binder &binder, CreateStatement &stmt, TableCatalogEntry &table,
 	                                            unique_ptr<LogicalOperator> plan) override;
-
 	DatabaseSize GetDatabaseSize(ClientContext &context) override;
-
-	IRCEndpointBuilder GetBaseUrl() const;
-
 	//! Whether or not this is an in-memory Iceberg database
 	bool InMemory() override;
 	string GetDBPath() override;
 
-	void ClearCache();
-
-	string OptionalGetCachedValue(const string &url);
-	bool SetCachedValue(const string &url, const string &value, const rest_api_objects::LoadTableResult &result);
-
-private:
-	void DropSchema(ClientContext &context, DropInfo &info) override;
+public:
+	AccessMode access_mode;
+	unique_ptr<IRCAuthorization> auth_handler;
+	IRCEndpointBuilder endpoint_builder;
+	//! warehouse
+	string warehouse;
+	//! host of the REST catalog
+	string uri;
+	//! version
+	const string version;
+	//! optional prefix
+	string prefix;
 
 private:
 	IRCSchemaSet schemas;
-	string default_schema;
 
 	// defaults and overrides provided by a catalog.
 	case_insensitive_map_t<string> defaults;
