+++ src/storage/irc_schema_set.cpp
@@ -34,7 +34,7 @@ void IRCSchemaSet::FillEntry(ClientContext &context, unique_ptr<CatalogEntry> &e
 
 optional_ptr<CatalogEntry> IRCSchemaSet::CreateSchema(ClientContext &context, CreateSchemaInfo &info) {
 	auto &ic_catalog = catalog.Cast<IRCatalog>();
-	auto schema = IRCAPI::CreateSchema(context, ic_catalog, ic_catalog.internal_name, info.schema);
+	auto schema = IRCAPI::CreateSchema(context, ic_catalog, info.schema);
 	auto schema_entry = make_uniq<IRCSchemaEntry>(catalog, info);
 	schema_entry->schema_data = make_uniq<IRCAPISchema>(schema);
 	return CreateEntry(std::move(schema_entry));
@@ -42,7 +42,7 @@ optional_ptr<CatalogEntry> IRCSchemaSet::CreateSchema(ClientContext &context, Cr
 
 void IRCSchemaSet::DropSchema(ClientContext &context, DropInfo &info) {
 	auto &ic_catalog = catalog.Cast<IRCatalog>();
-	IRCAPI::DropSchema(context, ic_catalog.internal_name, info.name);
+	IRCAPI::DropSchema(context, info.name);
 	DropEntry(context, info);
 }
